// monorepo/examples/web.chat/proto/gen/devinternal_pb.ts
export enum Tags {
  /**
   * @generated from enum value: UNKNOWN_TAG = 0;
   */
  UNKNOWN_TAG = 0,

  /**
   * ScopeTypes
   *
   * @generated from enum value: SCOPE_TYPE = 10;
   */
  SCOPE_TYPE = 10,

  /**
   * PathTypes
   *
   * @generated from enum value: PATH_TYPE = 20;
   */
  PATH_TYPE = 20,

  /**
   * @generated from enum value: APP_ID = 105;
   */
  APP_ID = 105,

  /**
   * @generated from enum value: APP_CHANNEL_ID = 110;
   */
  APP_CHANNEL_ID = 110,

  /**
   * @generated from enum value: APP_GROUP_ID = 115;
   */
  APP_GROUP_ID = 115,

  /**
   * @generated from enum value: APP_SUB_GROUP_ID = 120;
   */
  APP_SUB_GROUP_ID = 120,

  /**
   * @generated from enum value: APP_ROOM_ID = 125;
   */
  APP_ROOM_ID = 125,

  /**
   * @generated from enum value: APP_USER_ID = 130;
   */
  APP_USER_ID = 130,

  /**
   * @generated from enum value: APP_SESSION_ID = 135;
   */
  APP_SESSION_ID = 135,

  /**
   * @generated from enum value: APP_BLAH_BLAH_ID = 140;
   */
  APP_BLAH_BLAH_ID = 140,

  /**
   * @generated from enum value: APP_HANDLE_RAISED = 145;
   */
  APP_HANDLE_RAISED = 145,

  /**
   * @generated from enum value: DB_NAME = 250;
   */
  DB_NAME = 250,

  /**
   * @generated from enum value: DB_TABLE = 255;
   */
  DB_TABLE = 255,

  /**
   * @generated from enum value: DB_ROW_ID = 260;
   */
  DB_ROW_ID = 260,

  /**
   * @generated from enum value: DB_COLUMN_ID = 265;
   */
  DB_COLUMN_ID = 265,

  /**
   * @generated from enum value: DB_BLAH_BLAH_ID = 270;
   */
  DB_BLAH_BLAH_ID = 270,
}
// ...
export enum ScopeTypes {
  /**
   * @generated from enum value: UNKNOWN_SCOPE_TYPE = 0;
   */
  UNKNOWN_SCOPE_TYPE = 0,

  /**
   * @generated from enum value: SYSTEM = 10;
   */
  SYSTEM = 10,

  /**
   * @generated from enum value: USER = 20;
   */
  USER = 20,
}
// Retrieve enum metadata with: proto3.getEnumType(ScopeTypes)
proto3.util.setEnumType(ScopeTypes, "aircore.media.omni.v1.ScopeTypes", [
  { no: 0, name: "UNKNOWN_SCOPE_TYPE" },
  { no: 10, name: "SYSTEM" },
  { no: 20, name: "USER" },
]);

/**
 * ie PATH_TYPE tag, specific to a sequence_number_path format
 *
 * @generated from enum aircore.media.omni.v1.PathTypes
 */
export enum PathTypes {
  /**
   * @generated from enum value: UNKNOWN_PATH_TYPE = 0;
   */
  UNKNOWN_PATH_TYPE = 0,

  /**
   * @generated from enum value: SEQ_APP = 5;
   */
  SEQ_APP = 5,

  /**
   * @generated from enum value: SEQ_APP_CHAN = 10;
   */
  SEQ_APP_CHAN = 10,

  /**
   * @generated from enum value: SEQ_APP_ROOM = 15;
   */
  SEQ_APP_ROOM = 15,

  /**
   * @generated from enum value: SEQ_APP_ROOM_GROUP = 20;
   */
  SEQ_APP_ROOM_GROUP = 20,

  /**
   * @generated from enum value: ITEM_APP_CHAN_USER = 25;
   */
  ITEM_APP_CHAN_USER = 25,

  /**
   * @generated from enum value: SEQ_DB_NAME_TABLE = 30;
   */
  SEQ_DB_NAME_TABLE = 30,

  /**
   * @generated from enum value: ITEM_DB_NAME_TABLE_ROW_COLUMN = 50;
   */
  ITEM_DB_NAME_TABLE_ROW_COLUMN = 50,

  /**
   * @generated from enum value: SEQ_RECORDING = 60;
   */
  SEQ_RECORDING = 60,

  /**
   * @generated from enum value: ITEM_RECORDING = 65;
   */
  ITEM_RECORDING = 65,
}
// ...
export class PathElement extends Message<PathElement> {
  /**
   * @generated from field: aircore.media.omni.v1.Tags tag = 10;
   */
  tag = Tags.UNKNOWN_TAG;

  /**
   * @generated from oneof aircore.media.omni.v1.PathElement.x
   */
  x: {
    /**
     * @generated from field: aircore.media.omni.v1.ScopeTypes scope_type = 15;
     */
    value: ScopeTypes;
    case: "scopeType";
  } | {
    /**
     * @generated from field: aircore.media.omni.v1.PathTypes path_type = 20;
     */
    value: PathTypes;
    case: "pathType";
  } | {
    /**
     * @generated from field: string text = 25;
     */
    value: string;
    case: "text";
  } | {
    /**
     * @generated from field: google.protobuf.Any proto = 30;
     */
    value: Any;
    case: "proto";
  } | {
    /**
     * @generated from field: bytes buffer = 35;
     */
    value: Uint8Array;
    case: "buffer";
  } | {
    /**
     * @generated from field: int32 integer = 50;
     */
    value: number;
    case: "integer";
  } | {
    /**
     * @generated from field: float fraction = 55;
     */
    value: number;
    case: "fraction";
  } | { case: undefined; value?: undefined } = { case: undefined };

  constructor(data?: PartialMessage<PathElement>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "aircore.media.omni.v1.PathElement";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 10, name: "tag", kind: "enum", T: proto3.getEnumType(Tags) },
    { no: 15, name: "scope_type", kind: "enum", T: proto3.getEnumType(ScopeTypes), oneof: "x" },
    { no: 20, name: "path_type", kind: "enum", T: proto3.getEnumType(PathTypes), oneof: "x" },
    { no: 25, name: "text", kind: "scalar", T: 9 /* ScalarType.STRING */, oneof: "x" },
    { no: 30, name: "proto", kind: "message", T: Any, oneof: "x" },
    { no: 35, name: "buffer", kind: "scalar", T: 12 /* ScalarType.BYTES */, oneof: "x" },
    { no: 50, name: "integer", kind: "scalar", T: 5 /* ScalarType.INT32 */, oneof: "x" },
    { no: 55, name: "fraction", kind: "scalar", T: 2 /* ScalarType.FLOAT */, oneof: "x" },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): PathElement {
    return new PathElement().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): PathElement {
    return new PathElement().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): PathElement {
    return new PathElement().fromJsonString(jsonString, options);
  }

  static equals(a: PathElement | PlainMessage<PathElement> | undefined, b: PathElement | PlainMessage<PathElement> | undefined): boolean {
    return proto3.util.equals(PathElement, a, b);
  }
}

/**
 * @generated from message aircore.media.omni.v1.Path
 */
export class Path extends Message<Path> {
  /**
   * first hop is always a PATH_TYPE:PathTypes.xxx
   *
   * @generated from field: repeated aircore.media.omni.v1.PathElement hops = 10;
   */
  hops: PathElement[] = [];

  constructor(data?: PartialMessage<Path>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "aircore.media.omni.v1.Path";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 10, name: "hops", kind: "message", T: PathElement, repeated: true },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Path {
    return new Path().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Path {
    return new Path().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Path {
    return new Path().fromJsonString(jsonString, options);
  }

  static equals(a: Path | PlainMessage<Path> | undefined, b: Path | PlainMessage<Path> | undefined): boolean {
    return proto3.util.equals(Path, a, b);
  }
}
// ...
export class Sequencing extends Message<Sequencing> {
  /**
   * @generated from field: google.protobuf.Timestamp epoc = 10;
   */
  epoc?: Timestamp;

  /**
   * @generated from field: int64 sequence_number = 20;
   */
  sequenceNumber = protoInt64.zero;

  constructor(data?: PartialMessage<Sequencing>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "aircore.media.omni.v1.Sequencing";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 10, name: "epoc", kind: "message", T: Timestamp },
    { no: 20, name: "sequence_number", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Sequencing {
    return new Sequencing().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Sequencing {
    return new Sequencing().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Sequencing {
    return new Sequencing().fromJsonString(jsonString, options);
  }

  static equals(a: Sequencing | PlainMessage<Sequencing> | undefined, b: Sequencing | PlainMessage<Sequencing> | undefined): boolean {
    return proto3.util.equals(Sequencing, a, b);
  }
}

/**
 * @generated from message aircore.media.omni.v1.Payload
 */
export class Payload extends Message<Payload> {
  /**
   * @generated from oneof aircore.media.omni.v1.Payload.x
   */
  x: {
    /**
     * @generated from field: bytes buffer = 50;
     */
    value: Uint8Array;
    case: "buffer";
  } | {
    /**
     * @generated from field: string text = 60;
     */
    value: string;
    case: "text";
  } | {
    /**
     * @generated from field: google.protobuf.Any val = 70;
     */
    value: Any;
    case: "val";
  } | { case: undefined; value?: undefined } = { case: undefined };

  /**
   * @generated from field: aircore.media.omni.v1.Sequencing sequencing = 90;
   */
  sequencing?: Sequencing;

  /**
   * @generated from field: aircore.media.omni.v1.Path item_path = 100;
   */
  itemPath?: Path;

  constructor(data?: PartialMessage<Payload>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "aircore.media.omni.v1.Payload";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 50, name: "buffer", kind: "scalar", T: 12 /* ScalarType.BYTES */, oneof: "x" },
    { no: 60, name: "text", kind: "scalar", T: 9 /* ScalarType.STRING */, oneof: "x" },
    { no: 70, name: "val", kind: "message", T: Any, oneof: "x" },
    { no: 90, name: "sequencing", kind: "message", T: Sequencing },
    { no: 100, name: "item_path", kind: "message", T: Path },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): Payload {
    return new Payload().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): Payload {
    return new Payload().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): Payload {
    return new Payload().fromJsonString(jsonString, options);
  }

  static equals(a: Payload | PlainMessage<Payload> | undefined, b: Payload | PlainMessage<Payload> | undefined): boolean {
    return proto3.util.equals(Payload, a, b);
  }
}
// ...
export class UpsertRequest extends Message<UpsertRequest> {
  /**
   * @generated from field: aircore.media.omni.v1.Path sequence_number_path = 10;
   */
  sequenceNumberPath?: Path;

  /**
   * @generated from field: aircore.media.omni.v1.Payload payload = 30;
   */
  payload?: Payload;

  constructor(data?: PartialMessage<UpsertRequest>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "aircore.media.omni.v1.UpsertRequest";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 10, name: "sequence_number_path", kind: "message", T: Path },
    { no: 30, name: "payload", kind: "message", T: Payload },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): UpsertRequest {
    return new UpsertRequest().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): UpsertRequest {
    return new UpsertRequest().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): UpsertRequest {
    return new UpsertRequest().fromJsonString(jsonString, options);
  }

  static equals(a: UpsertRequest | PlainMessage<UpsertRequest> | undefined, b: UpsertRequest | PlainMessage<UpsertRequest> | undefined): boolean {
    return proto3.util.equals(UpsertRequest, a, b);
  }
}